+++ config/prod/hibernate.cfg.xml
@@ -5,7 +5,7 @@
 <hibernate-configuration>
     <session-factory>
         <property name="connection.url">
-            <![CDATA[jdbc:mysql://localhost:3306/tickets?autoReconnect=true&user=root&password=vlad123]]></property>
+            <![CDATA[jdbc:mysql://localhost:3306/tickets?autoReconnect=true&user=xxx&password=xxx]]></property>
         <property name="connection.provider_class">com.zaxxer.hikari.hibernate.HikariConnectionProvider
         </property>
         <property name="connection.driver_class">org.mariadb.jdbc.Driver</property>
@@ -20,12 +20,12 @@
         <property name="hibernate.current_session_context_class">thread</property>
         <property name="hibernate.connection.release_mode">after_transaction</property>
 
-        <property name="hibernate.cache.use_second_level_cache">false</property>
-        <property name="hibernate.cache.use_query_cache">false</property>
+        <property name="hibernate.cache.use_second_level_cache">true</property>
+        <property name="hibernate.cache.use_query_cache">true</property>
         <property name="hibernate.generate_statistics">false</property>
         <property name="hibernate.cache.use_structured_entries">true</property>
 
-        <property name="hibernate.show_sql">true</property>
+        <property name="hibernate.show_sql">false</property>
         <property name="hibernate.format_sql">false</property>
         <property name="hibernate.use_sql_comments">false</property>
         <property name="hibernate.enable_lazy_load_no_trans">true</property>
