+++ config/prod/log4j2.xml
@@ -19,15 +19,15 @@
     </Appenders>
 
     <Loggers>
-        <Root level="debug">
+        <Root level="info">
             <AppenderRef ref="RollingFile"/>
             <AppenderRef ref="STDOUT"/>
         </Root>
-        <Logger name="ro.igpr.tickets" level="debug"/>
-        <Logger name="org.restexpress.domain" level="debug"/>
-        <Logger name="org.hibernate" level="info"/>
-        <Logger name="org.hibernate.SQL" level="info"/>
-        <Logger name="org.hibernate.type" level="info"/>
+        <Logger name="ro.igpr.tickets" level="error"/>
+        <Logger name="org.restexpress.domain" level="error"/>
+        <Logger name="org.hibernate" level="error"/>
+        <Logger name="org.hibernate.SQL" level="error"/>
+        <Logger name="org.hibernate.type" level="error"/>
 
         <logger name="com.zaxxer.hikari" level="error"/>
         <Logger name="org.hibernate.hikaricp" level="error"/>
