+++ src/main/java/ro/igpr/tickets/domain/BaseEntity.java
@@ -1,0 +1,88 @@
+package ro.igpr.tickets.domain;
+
+import com.strategicgains.restexpress.plugin.swagger.annotations.ApiModelProperty;
+import org.hibernate.annotations.GenericGenerator;
+
+import javax.persistence.*;
+import java.util.Date;
+
+@MappedSuperclass
+public abstract class BaseEntity {
+
+    @ApiModelProperty(required = true)
+    protected Integer id;
+    protected Date createDate;
+    protected Date updateDate;
+    protected Date deleteDate;
+
+    @Id
+    @GeneratedValue(generator = "increment")
+    @GenericGenerator(name = "increment", strategy = "increment")
+    @Column(name = "`id`", nullable = false, insertable = true, updatable = false, unique = true)
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+    @Basic
+    @Temporal(TemporalType.DATE)
+    @Column(name = "`create_date`", nullable = true, insertable = true, updatable = false)
+    public Date getCreateDate() {
+        return createDate;
+    }
+
+    public void setCreateDate(Date createDate) {
+        this.createDate = createDate;
+    }
+
+    @Basic
+    @Temporal(TemporalType.DATE)
+    @Column(name = "`update_date`", nullable = true, insertable = true, updatable = true)
+    public Date getUpdateDate() {
+        return updateDate;
+    }
+
+    public void setUpdateDate(Date updateDate) {
+        this.updateDate = updateDate;
+    }
+
+
+    @Basic
+    @Temporal(TemporalType.DATE)
+    @Column(name = "`delete_date`", nullable = true, insertable = true, updatable = true)
+    public Date getDeleteDate() {
+        return deleteDate;
+    }
+
+    public void setDeleteDate(Date deleteDate) {
+        this.deleteDate = deleteDate;
+    }
+
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof BaseEntity)) return false;
+
+        BaseEntity that = (BaseEntity) o;
+
+        if (getId() != null ? !getId().equals(that.getId()) : that.getId() != null) return false;
+        if (getCreateDate() != null ? !getCreateDate().equals(that.getCreateDate()) : that.getCreateDate() != null)
+            return false;
+        if (getUpdateDate() != null ? !getUpdateDate().equals(that.getUpdateDate()) : that.getUpdateDate() != null)
+            return false;
+        return getDeleteDate() != null ? getDeleteDate().equals(that.getDeleteDate()) : that.getDeleteDate() == null;
+    }
+
+    @Override
+    public int hashCode() {
+        int result = getId() != null ? getId().hashCode() : 0;
+        result = 31 * result + (getCreateDate() != null ? getCreateDate().hashCode() : 0);
+        result = 31 * result + (getUpdateDate() != null ? getUpdateDate().hashCode() : 0);
+        result = 31 * result + (getDeleteDate() != null ? getDeleteDate().hashCode() : 0);
+        return result;
+    }
+}
